cv_service/requirements.txt
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
numpy==1.26.4
confluent-kafka==2.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
